+++ src/main/java/net/pwall/util/pipeline/AbstractAcceptor.java
@@ -60,14 +60,4 @@ abstract public class AbstractAcceptor<A, R> extends BaseAbstractAcceptor<R> imp
      */
     abstract public void acceptObject(A value) throws Exception;
 
-    /**
-     * Get the result object (default implementation throws exception).
-     *
-     * @throws  IllegalStateException   the acceptor does not have a result
-     */
-    @Override
-    public R getResult() {
-        throw new IllegalStateException("Acceptor does not have a result");
-    }
-
 }